websocket_server/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==1.1.2
#socketIO
#the flask-socketio import actually handles compatabilities for us
# as opposed to importing python-socketio and engineio of diff version
Flask-SocketIO==4.3.2
#messaging
aio-pika
#http
aiohttp==3.7.4.post0